requirements.txt
langchain==0.0.327
pydantic==2.4.2
python-dotenv==1.0.0
